// src/parity/utils/types.ts
export interface MoveRoute {
  list: MoveCommand[];
  repeat: boolean;
  skippable: boolean;
  wait: boolean;
}

export interface MoveCommand {
  code: number;
  parameters: unknown[];
}